+++ app/src/androidTest/java/woowacourse/shopping/ui/cart/CartActivityTest.kt
@@ -18,6 +18,7 @@ import org.junit.runner.RunWith
 import woowacourse.shopping.R
 import woowacourse.shopping.model.Product
 import woowacourse.shopping.model.data.CartsImpl
+import woowacourse.shopping.ui.cart.adapter.CartViewHolder
 
 @RunWith(AndroidJUnit4::class)
 class CartActivityTest {
