+++ app/src/androidTest/java/woowacourse/shopping/ui/products/ProductContentsActivityTest.kt
@@ -15,7 +15,8 @@ import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
 import woowacourse.shopping.R
-import woowacourse.shopping.ui.cart.CartViewHolder
+import woowacourse.shopping.ui.cart.adapter.CartViewHolder
+import woowacourse.shopping.ui.products.adapter.ProductViewHolder
 
 @RunWith(AndroidJUnit4::class)
 class ProductContentsActivityTest {
