+++ app/src/main/java/woowacourse/shopping/ui/cart/viewmodel/CartViewModel.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.ui.cart
+package woowacourse.shopping.ui.cart.viewmodel
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
