+++ app/src/main/java/woowacourse/shopping/ui/cart/viewmodel/CartViewModelFactory.kt
@@ -1,15 +1,13 @@
-package woowacourse.shopping.ui.detail
+package woowacourse.shopping.ui.cart.viewmodel
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import woowacourse.shopping.model.data.CartDao
-import woowacourse.shopping.model.data.ProductDao
 
-class ProductDetailViewModelFactory(
-    private val productDao: ProductDao,
+class CartViewModelFactory(
     private val cartDao: CartDao,
 ) : ViewModelProvider.Factory {
     override fun <T : ViewModel> create(modelClass: Class<T>): T {
-        return ProductDetailViewModel(productDao, cartDao) as T
+        return CartViewModel(cartDao) as T
     }
 }
