+++ app/src/main/java/woowacourse/shopping/ui/detail/viewmodel/ProductDetailViewModelFactory.kt
@@ -1,13 +1,15 @@
-package woowacourse.shopping.ui.cart
+package woowacourse.shopping.ui.detail.viewmodel
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import woowacourse.shopping.model.data.CartDao
+import woowacourse.shopping.model.data.ProductDao
 
-class CartViewModelFactory(
+class ProductDetailViewModelFactory(
+    private val productDao: ProductDao,
     private val cartDao: CartDao,
 ) : ViewModelProvider.Factory {
     override fun <T : ViewModel> create(modelClass: Class<T>): T {
-        return CartViewModel(cartDao) as T
+        return ProductDetailViewModel(productDao, cartDao) as T
     }
 }
