+++ app/src/main/java/woowacourse/shopping/ui/products/adapter/ProductAdapter.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.ui.products
+package woowacourse.shopping.ui.products.adapter
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
