+++ app/src/main/java/woowacourse/shopping/ui/products/viewmodel/ProductContentsViewModel.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.ui.products
+package woowacourse.shopping.ui.products.viewmodel
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
