+++ app/src/main/java/woowacourse/shopping/ui/products/viewmodel/ProductContentsViewModelFactory.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.ui.products
+package woowacourse.shopping.ui.products.viewmodel
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
