+++ app/src/main/res/layout/activity_cart.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="vm"
-            type="woowacourse.shopping.ui.cart.CartViewModel" />
+            type="woowacourse.shopping.ui.cart.viewmodel.CartViewModel" />
 
     </data>
