+++ app/src/test/java/woowacourse/shopping/ui/cart/CartViewModelTest.kt
@@ -8,6 +8,7 @@ import woowacourse.shopping.InstantTaskExecutorExtension
 import woowacourse.shopping.getOrAwaitValue
 import woowacourse.shopping.model.Product
 import woowacourse.shopping.model.data.CartsImpl
+import woowacourse.shopping.ui.cart.viewmodel.CartViewModel
 
 @ExtendWith(InstantTaskExecutorExtension::class)
 class CartViewModelTest {
