+++ app/src/test/java/woowacourse/shopping/ui/detail/ProductDetailViewModelTest.kt
@@ -10,6 +10,7 @@ import woowacourse.shopping.getOrAwaitValue
 import woowacourse.shopping.model.Product
 import woowacourse.shopping.model.data.CartsImpl
 import woowacourse.shopping.model.data.ProductsImpl
+import woowacourse.shopping.ui.detail.viewmodel.ProductDetailViewModel
 import woowacourse.shopping.ui.state.UiState
 
 @ExtendWith(InstantTaskExecutorExtension::class)
